+++ agent/lm_agent/backend_utils.py
@@ -149,8 +149,8 @@ async def get_license_manager_backend_version() -> str:
 class BackendConfigurationRow(BaseModel):
     """
     NOTE: This is a copy of the schema from the backend.
-          If the schema changes upstream in a non-reverse-compatible
-          way, this schema should cause errors in deserialization.
+          If the schema changes upstream in a non-reverse-compatible way, this schema should cause errors
+          in deserialization.
     """
 
     class Config: