+++ agent/lm_agent/exceptions.py
@@ -5,6 +5,10 @@ Custom exceptions for the License Manager Agent.
 from buzz import Buzz
 
 
+class LicenseManagerAuthTokenError(Buzz):
+    """Exception for backend connection issues."""
+
+
 class LicenseManagerBackendConnectionError(Buzz):
     """Exception for backend connection issues."""
 
@@ -13,5 +17,13 @@ class LicenseManagerBackendVersionError(Buzz):
     """Exception for backend/agent version mismatches."""
 
 
-class LicenseManagerAuthTokenError(Buzz):
-    """Exception for backend connection issues."""
+class LicenseManagerEmptyReportError(Buzz):
+    """Exception for empty report when no licenses added in backend."""
+
+
+class LicenseManagerNonSupportedServerTypeError(Buzz):
+    """Exception for entry with non supported server type."""
+
+
+class LicenseManagerBadServerOutput(Buzz):
+    """Exception for license server bad output."""