+++ agent/tests/test_tokenstat.py
@@ -1,49 +1,52 @@
-from pathlib import Path
 from textwrap import dedent
 from typing import List
 from unittest import mock
-from unittest.mock import patch
 
 from pytest import fixture, mark, raises
 
 from lm_agent import tokenstat
 from lm_agent.backend_utils import BackendConfigurationRow
-from lm_agent.parsing import flexlm, rlm
-from lm_agent.workload_managers.slurm import cmd_utils
-from tests.conftest import MOCK_BIN_PATH
+from lm_agent.config import settings
+from lm_agent.exceptions import LicenseManagerBadServerOutput
 
 
 @fixture
-def one_configuration_row():
-    return [
-        BackendConfigurationRow(
-            product="testproduct1",
-            features={"feature1": 10},
-            license_servers=["flexlm:127.0.0.1:2345"],
-            license_server_type="flexlm",
-            grace_time=10000,
-        )
-    ]
+def one_configuration_row_flexlm():
+    return BackendConfigurationRow(
+        product="testproduct",
+        features={"testfeature": 10},
+        license_servers=["flexlm:127.0.0.1:2345"],
+        license_server_type="flexlm",
+        grace_time=10000,
+    )
 
 
 @fixture
 def one_configuration_row_rlm():
-    return [
-        BackendConfigurationRow(
-            product="converge",
-            features={"super": 10},
-            license_servers=["rlm:127.0.0.1:2345"],
-            license_server_type="rlm",
-            grace_time=10000,
-        )
-    ]
+    return BackendConfigurationRow(
+        product="converge",
+        features={"super": 10},
+        license_servers=["rlm:127.0.0.1:2345"],
+        license_server_type="rlm",
+        grace_time=10000,
+    )
 
 
 @fixture
-def scontrol_show_lic_output():
+def flexlm_server(one_configuration_row_flexlm):
+    return tokenstat.FlexLMLicenseServer(one_configuration_row_flexlm.license_servers)
+
+
+@fixture
+def rlm_server(one_configuration_row_rlm):
+    return tokenstat.RLMLicenseServer(one_configuration_row_rlm.license_servers)
+
+
+@fixture
+def scontrol_show_lic_output_flexlm():
     return dedent(
         """
-        LicenseName=testproduct1.feature1@flexlm
+        LicenseName=testproduct.testfeature@flexlm
             Total=10 Used=0 Free=10 Reserved=0 Remote=yes
         """
     )
@@ -59,61 +62,84 @@ def scontrol_show_lic_output_rlm():
     )
 
 
-@fixture
-def license_server_features():
+def test_get_flexlm_commands_list(flexlm_server):
     """
-    The license server type, product and features.
+    Do the commands for invoking the license server have the correct data?
     """
-    return [{"features": ["TESTFEATURE"], "license_server_type": "flexlm", "product": "TESTPRODUCT"}]
+    commands_list = flexlm_server.get_commands_list()
+    assert commands_list == [f"{settings.LMUTIL_PATH} lmstat -c 2345@127.0.0.1 -f"]
 
 
-@fixture
-def license_server_features_rlm():
+def test_get_rlm_commands_list(rlm_server):
     """
-    The license server type, product and features.
+    Do the commands for invoking the license server have the correct data?
     """
-    return [{"features": ["super"], "license_server_type": "rlm", "product": "converge"}]
+    commands_list = rlm_server.get_commands_list()
+    assert commands_list == [f"{settings.RLMUTIL_PATH} rlmstat -c 2345@127.0.0.1 -a -p"]
 
 
-@fixture
-def tool_opts() -> tokenstat.ToolOptions:
+@mark.asyncio
+@mock.patch("lm_agent.tokenstat.asyncio.wait_for")
+@mock.patch("lm_agent.tokenstat.asyncio.create_subprocess_shell")
+async def test_flexlm_get_output_from_server(
+    create_subprocess_mock: mock.MagicMock,
+    wait_for_mock: mock.MagicMock,
+    flexlm_server,
+    lmstat_output,
+):
     """
-    A ToolOptions set up for easy testing
+    Do the license server interface return the output from the license server?
     """
-    return tokenstat.ToolOptions(
-        name="flexlm",
-        path=Path(f"{MOCK_BIN_PATH}/lmstat"),
-        args="{exe} {host} {port}",
-        parse_fn=flexlm.parse,
+    proc_mock = mock.MagicMock()
+    proc_mock.returncode = 0
+
+    create_subprocess_mock.return_value = proc_mock
+    wait_for_mock.return_value = (
+        bytes(lmstat_output, encoding="UTF8"),
+        None,
     )
 
+    output = await flexlm_server.get_output_from_server("testproduct.testfeature")
+    assert output == lmstat_output
 
-@fixture
-def tool_opts_rlm() -> tokenstat.ToolOptions:
+
+@mark.asyncio
+@mock.patch("lm_agent.tokenstat.asyncio.wait_for")
+@mock.patch("lm_agent.tokenstat.asyncio.create_subprocess_shell")
+async def test_rlm_get_output_from_server(
+    create_subprocess_mock: mock.MagicMock,
+    wait_for_mock: mock.MagicMock,
+    rlm_server,
+    rlm_output,
+):
     """
-    A ToolOptions set up for easy testing
+    Do the license server interface return the output from the license server?
     """
-    return tokenstat.ToolOptions(
-        name="rlm",
-        path=MOCK_BIN_PATH / "rlmstat",
-        args="{exe} {host} {port}",
-        parse_fn=rlm.parse,
+    proc_mock = mock.MagicMock()
+    proc_mock.returncode = 0
+
+    create_subprocess_mock.return_value = proc_mock
+    wait_for_mock.return_value = (
+        bytes(rlm_output, encoding="UTF8"),
+        None,
     )
 
+    output = await rlm_server.get_output_from_server()
+    assert output == rlm_output
 
-def test_lri_from_stdout(lm_output):
+
+@mark.asyncio
+@mock.patch("lm_agent.tokenstat.FlexLMLicenseServer.get_output_from_server")
+async def test_flexlm_get_report_item(
+    get_output_from_server_mock: mock.MagicMock, flexlm_server, lmstat_output
+):
     """
-    Do I parse the stdout string from flexlm to produce structured data?
+    Do the FlexLM server interface generate a report item for the product?
     """
-    lri = tokenstat.LicenseReportItem.from_stdout(
-        product="TESTPRODUCT",
-        parse_fn=flexlm.parse,
-        tool_name="flexlm",
-        stdout=lm_output,
-    )
-    assert lri == tokenstat.LicenseReportItem(
-        tool_name="flexlm",
-        product_feature="TESTPRODUCT.TESTFEATURE",
+    get_output_from_server_mock.return_value = lmstat_output
+
+    assert await flexlm_server.get_report_item("testproduct.testfeature") == tokenstat.LicenseReportItem(
+        product_feature="testproduct.testfeature",
         used=93,
         total=1000,
         used_licenses=[
@@ -124,126 +150,124 @@ def test_lri_from_stdout(lm_output):
     )
 
 
-def test_tooloptions_cmd_list(tool_opts: tokenstat.ToolOptions, one_configuration_row):
-    """
-    Do I produce a reasonable set of command lines from the tool definitions
-    """
-    assert tool_opts.cmd_list(one_configuration_row[0].license_servers) == [
-        f"{MOCK_BIN_PATH}/lmstat 127.0.0.1 2345",
-    ]
-
-
 @mark.asyncio
-async def test_attempt_tool_checks(
-    tool_opts: tokenstat.ToolOptions,
-):
+@mock.patch("lm_agent.tokenstat.RLMLicenseServer.get_output_from_server")
+async def test_rlm_get_report_item(get_output_from_server_mock: mock.MagicMock, rlm_server, rlm_output):
     """
-    Do I run the commands corresponding to my collection of tools and service definitions?
+    Do the RLM server interface generate a report item for the product?
     """
-    # good tool
+    get_output_from_server_mock.return_value = rlm_output
 
-    ret = await tokenstat.attempt_tool_checks(
-        tool_opts,
-        "TESTPRODUCT",
-        "TESTFEATURE",
-        ["flexlm:127.0.0.1:2345"],
-    )
-    assert ret == tokenstat.LicenseReportItem(
-        tool_name="flexlm",
-        product_feature="TESTPRODUCT.TESTFEATURE",
-        used=502,
+    assert await rlm_server.get_report_item("converge.super") == tokenstat.LicenseReportItem(
+        product_feature="converge.super",
+        used=93,
         total=1000,
         used_licenses=[
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 29},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 27},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 23},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 9},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 8},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 15},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 43},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 11},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 13},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 23},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 28},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 11},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 17},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 38},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 25},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 35},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 38},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 4},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 9},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 11},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 19},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 15},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 14},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 37},
+            {"booked": 29, "user_name": "jbemfv", "lead_host": "myserver.example.com"},
+            {"booked": 27, "user_name": "cdxfdn", "lead_host": "myserver.example.com"},
+            {"booked": 37, "user_name": "jbemfv", "lead_host": "myserver.example.com"},
         ],
     )
 
-    # bad tool
-    tool_opts.args = "{exe} fail"
-    with raises(RuntimeError):
-        await tokenstat.attempt_tool_checks(
-            tool_opts,
-            "TESTPRODUCT",
-            "TESTFEATURE",
-            ["flexlm:127.0.0.1:2345"],
-        )
+
+@mark.asyncio
+@mock.patch("lm_agent.tokenstat.FlexLMLicenseServer.get_output_from_server")
+async def test_flexlm_get_report_item_with_bad_output(
+    get_output_from_server_mock: mock.MagicMock, flexlm_server, lmstat_output_bad
+):
+    get_output_from_server_mock.return_value = lmstat_output_bad
+
+    with raises(LicenseManagerBadServerOutput):
+        await flexlm_server.get_report_item("testproduct.testfeature")
 
 
 @mark.asyncio
-@mock.patch("lm_agent.tokenstat.get_config_from_backend")
+@mock.patch("lm_agent.tokenstat.RLMLicenseServer.get_output_from_server")
+async def test_rlm_get_report_item_with_no_used_licenses(
+    get_output_from_server_mock: mock.MagicMock, rlm_server, rlm_output_no_licenses
+):
+    get_output_from_server_mock.return_value = rlm_output_no_licenses
+
+    assert await rlm_server.get_report_item("converge.super") == tokenstat.LicenseReportItem(
+        product_feature="converge.super",
+        used=0,
+        total=1000,
+        used_licenses=[],
+    )
+
+
+@mark.asyncio
+@mock.patch("lm_agent.tokenstat.FlexLMLicenseServer.get_output_from_server")
+async def test_flexlm_get_report_item_with_no_used_licenses(
+    get_output_from_server_mock: mock.MagicMock, flexlm_server, lmstat_output_no_licenses
+):
+    get_output_from_server_mock.return_value = lmstat_output_no_licenses
+
+    assert await flexlm_server.get_report_item("testproduct.testfeature") == tokenstat.LicenseReportItem(
+        product_feature="testproduct.testfeature",
+        used=0,
+        total=1000,
+        used_licenses=[],
+    )
+
+
+@mark.asyncio
+@mark.parametrize(
+    "output,reconciliation",
+    [
+        (
+            "lmstat_output",
+            [
+                {
+                    "product_feature": "testproduct.testfeature",
+                    "used": 93,
+                    "total": 1000,
+                    "used_licenses": [
+                        {
+                            "user_name": "jbemfv",
+                            "lead_host": "myserver.example.com",
+                            "booked": 29,
+                        },
+                        {
+                            "user_name": "cdxfdn",
+                            "lead_host": "myserver.example.com",
+                            "booked": 27,
+                        },
+                        {
+                            "user_name": "jbemfv",
+                            "lead_host": "myserver.example.com",
+                            "booked": 37,
+                        },
+                    ],
+                },
+            ],
+        ),
+    ],
+)
+@mock.patch("lm_agent.tokenstat.FlexLMLicenseServer.get_output_from_server")
 @mock.patch("lm_agent.tokenstat.scontrol_show_lic")
-async def test_report(
-    show_lic_mock: mock.MagicMock,
+@mock.patch("lm_agent.tokenstat.get_config_from_backend")
+async def test_flexlm_get_report(
     get_config_from_backend_mock: mock.MagicMock,
-    tool_opts: tokenstat.ToolOptions,
-    one_configuration_row,
-    scontrol_show_lic_output,
+    show_lic_mock: mock.MagicMock,
+    get_output_from_server_mock: mock.MagicMock,
+    output,
+    reconciliation,
+    one_configuration_row_flexlm,
+    scontrol_show_lic_output_flexlm,
+    request,
 ):
     """
-    Do I collect the requested structured data from running all these dang tools?
+    Do I get a report for the FlexLM licenses in the cluster?
     """
-    get_config_from_backend_mock.return_value = one_configuration_row
-    show_lic_mock.return_value = scontrol_show_lic_output
+    get_config_from_backend_mock.return_value = [one_configuration_row_flexlm]
+    show_lic_mock.return_value = scontrol_show_lic_output_flexlm
 
-    # Patch the objects needed to generate a report.
-    p0 = patch.object(cmd_utils, "get_tokens_for_license", 0)
-    p1 = patch.dict(tokenstat.ToolOptionsCollection.tools, {"flexlm": tool_opts})
-    license_report_item = {
-        "product_feature": "testproduct1.TESTFEATURE",
-        "used": 502,
-        "total": 1000,
-        "used_licenses": [
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 29},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 27},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 23},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 9},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 8},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 15},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 43},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 11},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 13},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 23},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 28},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 11},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 17},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 38},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 25},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 35},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 38},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 4},
-            {"user_name": "jxezha", "lead_host": "myserver.example.com", "booked": 9},
-            {"user_name": "cdxfdn", "lead_host": "myserver.example.com", "booked": 11},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 19},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 15},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 14},
-            {"user_name": "jbemfv", "lead_host": "myserver.example.com", "booked": 37},
-        ],
-    }
-    with p0, p1:
-        assert [license_report_item] == await tokenstat.report()
+    output = request.getfixturevalue(output)
+    get_output_from_server_mock.return_value = output
+
+    reconcile_list = await tokenstat.report()
+    assert reconcile_list == reconciliation
 
 
 @mark.asyncio
@@ -290,40 +314,28 @@ async def test_report(
         ),
     ],
 )
+@mock.patch("lm_agent.tokenstat.RLMLicenseServer.get_output_from_server")
 @mock.patch("lm_agent.tokenstat.scontrol_show_lic")
 @mock.patch("lm_agent.tokenstat.get_config_from_backend")
-@mock.patch("lm_agent.tokenstat.asyncio.create_subprocess_shell")
-@mock.patch("lm_agent.tokenstat.asyncio.wait_for")
-@mock.patch("lm_agent.tokenstat.ToolOptionsCollection")
-async def test_report_rlm(
-    tools_mock: mock.MagicMock,
-    wait_for_mock: mock.AsyncMock,
-    create_subprocess_mock: mock.AsyncMock,
+async def test_rlm_get_report(
     get_config_from_backend_mock: mock.MagicMock,
     show_lic_mock: mock.MagicMock,
+    get_output_from_server_mock: mock.MagicMock,
     output,
     reconciliation,
-    tool_opts_rlm: tokenstat.ToolOptions,
     one_configuration_row_rlm,
     scontrol_show_lic_output_rlm,
     request,
 ):
     """
-    Do I collect the requested structured data from running all these dang tools?
+    Do I get a report for the RLM licenses in the cluster?
     """
-    proc_mock = mock.MagicMock()
-    proc_mock.returncode = 0
-    create_subprocess_mock.return_value = proc_mock
-    get_config_from_backend_mock.return_value = one_configuration_row_rlm
+    get_config_from_backend_mock.return_value = [one_configuration_row_rlm]
     show_lic_mock.return_value = scontrol_show_lic_output_rlm
 
-    tools_mock.tools = {"rlm": tool_opts_rlm}
     output = request.getfixturevalue(output)
+    get_output_from_server_mock.return_value = output
 
-    wait_for_mock.return_value = (
-        bytes(output, encoding="UTF8"),
-        None,
-    )
     reconcile_list = await tokenstat.report()
     assert reconcile_list == reconciliation
 
@@ -331,26 +343,34 @@ async def test_report_rlm(
 @mark.asyncio
 @mock.patch("lm_agent.tokenstat.scontrol_show_lic")
 @mock.patch("lm_agent.tokenstat.get_config_from_backend")
-@mock.patch("lm_agent.tokenstat.asyncio.create_subprocess_shell")
-@mock.patch("lm_agent.tokenstat.ToolOptionsCollection")
-async def test_report_rlm_empty_backend(
-    tools_mock: mock.MagicMock,
-    create_subprocess_mock: mock.AsyncMock,
+async def test_flexlm_report_with_empty_backend(
     get_config_from_backend_mock: mock.MagicMock,
     show_lic_mock: mock.MagicMock,
-    tool_opts_rlm: tokenstat.ToolOptions,
+    scontrol_show_lic_output_flexlm,
+):
+    """
+    Do I collect the requested structured data when the backend is empty?
+    """
+    get_config_from_backend_mock.return_value = []
+    show_lic_mock.return_value = scontrol_show_lic_output_flexlm
+
+    reconcile_list = await tokenstat.report()
+    assert reconcile_list == []
+
+
+@mark.asyncio
+@mock.patch("lm_agent.tokenstat.scontrol_show_lic")
+@mock.patch("lm_agent.tokenstat.get_config_from_backend")
+async def test_rlm_report_with_empty_backend(
+    get_config_from_backend_mock: mock.MagicMock,
+    show_lic_mock: mock.MagicMock,
     scontrol_show_lic_output_rlm,
 ):
     """
     Do I collect the requested structured data when the backend is empty?
     """
-    proc_mock = mock.MagicMock()
-    proc_mock.returncode = 0
-    create_subprocess_mock.return_value = proc_mock
     get_config_from_backend_mock.return_value = []
     show_lic_mock.return_value = scontrol_show_lic_output_rlm
-
-    tools_mock.tools = {"rlm": tool_opts_rlm}
 
     reconcile_list = await tokenstat.report()
     assert reconcile_list == []